Finished Spider Enemy
Created 2 types of the spider enemy, one that when the player gets close and underneath spider, it moves down and stops just above the ground, and another that after moving down, it stays for 3 seconds (can be changed)  and then raises to its original spot. The first is purple, and the second is orange. May need to change some stuff like the trigger for when the player gets close once I have access to the sprites for that, but overall its finished.