Major Camera Update
Finally finished the camera movement script. In order to use this script, remove the main camera from the Player object and make it its own object. Then add the script onto the camera. Or you can get rid of the camera entirely and just place a new one from the Camera Prefab. I removed the camera from the Player Prefab for obvious reasons.

I also created a camera prefab with default settings, along with a LeftBarrier prefab.

Features of Camera:
Smooth camera movement relative to the players position.
Able to limit where camera can go with variables.
Able to add offset from player so the player isn't always in the middle of the screen.
The camera can not move left.
Player can not go left of the camera using "LeftBarrier" Prefab (Place LeftBarrier anywhere in the scene and it will work).
Place camera anywhere in the scene and it will automatically go to players position or desired start position when run.

You can see how the camera is used in my test scene (jlgorcz Test Scene)