Added OneWay Platforms
Created two new one way objects. The first is just a block you place down from the prefab menu, and the other being a sperate tile set to make placing down for the level design team easy. There are some bugs with the tile set, and you can mess around with it in my Test Scene (jlgorcz Test Scene)
Also fixed a bug with the camera and optimized the camera prefab