Glitch Gun
The Glitch Gun power-up has been finished. When the glitch gun power-up is equipped, the bullet will travel for a editable amount of time before changing direction. There is a Boolean on the prefab whether after the bullet changes it should match its starting speed. While this makes the most sense for it to be toggled, I find it fun having a bullet accelerate infinitely after turning around.